+++ filters/zlibfilter.hpp
@@ -12,7 +12,6 @@ public:
   ~zlibFilter();
   void encode(File *in, File *out, uint64_t size, uint64_t info);
   uint64_t decode(File *in, File *out,  uint64_t size, uint64_t info);
-  
 };
 
 int parse_zlib_header(int header);
